final review, all modules modified
